requirements-render.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow>=10.0.0
huggingface_hub==0.36.0
requests==2.32.5
